requirements.txt
# Docker deployment requirements
pdf2image==1.17.0
pillow==11.3.0
flask
werkzeug
